ft-trainees:dosplaying trainees details
