+++ src/Mutations/manageStudentMutations.tsx
@@ -1,4 +1,5 @@
 import { gql } from '@apollo/client';
+import GET_PROFILE from './User';
 
 export const GET_USERS_QUERY = gql`
   query GetUsers($orgToken: String) {
@@ -13,7 +14,15 @@ export const GET_TRAINEES_QUERY = gql`
   query GetTrainees($orgToken: String) {
     getTrainees(orgToken: $orgToken) {
       profile {
-        name
+      firstName
+      lastName
+      city
+      country
+      phoneNumber
+      biography
+      avatar
+      id
+      name
       }
       email
       cohort {
@@ -34,6 +43,22 @@ export const GET_COHORTS_QUERY = gql`
     }
   }
 `;
+
+export const GET_TRAINEE_PROFILE = gql`
+query GetProfile {
+  getProfile {
+    firstName
+    name
+    city
+    country
+    address
+    phoneNumber
+    biography
+    avatar
+    coverImage
+    lastName
+  }
+}`;
 
 export const ADD_MEMBER_TO_COHORT_MUTATION = gql`
   mutation AddMemberToCohort(
